+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# 0.1.1
+- Add input validation for empty array.
+- Add input validation for non-Numo::NArray object.
+- Fix to raise NoMemoryError when happened memory allocation error.
+
 # 0.1.0
 - First release.